move make_load body into another file to keep lib.rs clean
